Alpha version of substitute_variables
Also moved name arg of linear_constraint and quadratic_constraint in the last
